Handling back button when the delegate have been removing

+++ DigitalLearningSolutions.Web/Controllers/TrackingSystem/Delegates/ActivityDelegatesController.cs
@@ -26,6 +26,7 @@
     using System;
     using System.Collections.Generic;
     using System.Linq;
+    using System.Net;
 
     [FeatureGate(FeatureFlags.RefactoredTrackingSystem)]
     [Authorize(Policy = CustomPolicies.UserCentreAdmin)]
@@ -354,6 +355,12 @@
         [HttpPost]
         public IActionResult RemoveDelegateSelfAssessment(DelegateSelfAssessmenteViewModel delegateSelfAssessmenteViewModel)
         {
+            var selfAssessmentDelegate = selfAssessmentService.CheckDelegateSelfAssessment(delegateSelfAssessmenteViewModel.CandidateAssessmentsId);
+
+            if (selfAssessmentDelegate > 0)
+            {
+                return StatusCode((int)HttpStatusCode.Gone);
+            }
             if (ModelState.IsValid && delegateSelfAssessmenteViewModel.ActionConfirmed)
             {
                 selfAssessmentService.RemoveDelegateSelfAssessment(delegateSelfAssessmenteViewModel.CandidateAssessmentsId);
